Fixed some issues with non-UTF8 characters being used from the XML file
git-svn-id: svn://svn.code.sf.net/p/open-dis/code@689 1da32bc5-2e4a-4c7f-92e1-9780f3f753aa

+++ languages/java/trunk/src/main/edu/nps/moves/dis/AngularVelocityVector.java
@@ -31,13 +31,13 @@ public class AngularVelocityVector extends Object implements Serializable
    private long pk_AngularVelocityVector;
 
    /** velocity about the x axis */
-   protected float  x = 0;
+   protected float  x = (float)0;
 
    /** velocity about the y axis */
-   protected float  y = 0;
+   protected float  y = (float)0;
 
    /** velocity about the zaxis */
-   protected float  z = 0;
+   protected float  z = (float)0;
 
 
 /** Constructor */
