+++ languages/java/trunk/src/main/edu/nps/moves/dis/FastEntityStatePdu.java
@@ -782,7 +782,7 @@ public void marshal(java.nio.ByteBuffer buff)
 
        for(int idx = 0; idx < otherParameters.length; idx++)
        {
-           buff.put(otherParameters[idx]);
+           buff.put((byte)otherParameters[idx]);
        } // end of array marshaling
 
        buff.putFloat( (float)xAcceleration);
@@ -794,7 +794,7 @@ public void marshal(java.nio.ByteBuffer buff)
 
        for(int idx = 0; idx < marking.length; idx++)
        {
-           buff.put(marking[idx]);
+           buff.put((byte)marking[idx]);
        } // end of array marshaling
 
        buff.putInt( (int)capabilities);
