+++ languages/java/trunk/src/main/edu/nps/moves/dismobile/AngularVelocityVector.java
@@ -17,13 +17,13 @@ import edu.nps.moves.disutil.*;
 public class AngularVelocityVector extends Object implements Serializable
 {
    /** velocity about the x axis */
-   protected float  x = 0;
+   protected float  x = (float)0;
 
    /** velocity about the y axis */
-   protected float  y = 0;
+   protected float  y = (float)0;
 
    /** velocity about the zaxis */
-   protected float  z = 0;
+   protected float  z = (float)0;
 
 
 /** Constructor */
