+++ languages/java/trunk/src/main/edu/nps/moves/dismobile/DataPdu.java
@@ -20,7 +20,7 @@ public class DataPdu extends SimulationManagementFamilyPdu implements Serializab
    protected long  requestID;
 
    /** padding */
-   protected long  padding1 = 0;
+   protected long  padding1 = (long)0;
 
    /** Number of fixed datum records */
    protected long  numberOfFixedDatumRecords;
