+++ languages/java/trunk/src/main/edu/nps/moves/dismobile/ResupplyReceivedPdu.java
@@ -26,10 +26,10 @@ public class ResupplyReceivedPdu extends LogisticsFamilyPdu implements Serializa
    protected short  numberOfSupplyTypes;
 
    /** padding */
-   protected short  padding1 = 0;
+   protected short  padding1 = (short)0;
 
    /** padding */
-   protected byte  padding2 = 0;
+   protected byte  padding2 = (byte)0;
 
    protected List< SupplyQuantity> supplies = new ArrayList<SupplyQuantity>();
